Add handler test struct to internal/test
